Fix edge cases, UB, integer overflows, and buffer state consistency
- Fix signed integer shift UB in GetCPUFeatures() by using 1u << i.
- Guard against duplicate '%' character in GetBatteryCapacity().
- Guard against unsigned underflow in GetSwapUsage() when free > max.
- Cast block counts and page counts to uint64 before multiplication/addition in GetRootDiskUsage(), GetDiskInfo(), and _GetCachedUsage().
- Reset fSize = 0 and preserve fInitStatus when CircularBuffer::operator= copies from an unallocated buffer.
- Expand test suite in test_circular_buffer.cpp and test_utils.cpp to cover all fixed edge cases.

+++ SystemMonitor/SystemDetailsView.cpp
@@ -413,7 +413,7 @@ BString SystemDetailsView::_GetRamUsage(system_info* sysInfo)
 
 BString SystemDetailsView::_GetCachedUsage(system_info* sysInfo)
 {
-	uint64 cachedBytes = static_cast<uint64>(sysInfo->cached_pages + sysInfo->block_cache_pages) * B_PAGE_SIZE;
+	uint64 cachedBytes = (static_cast<uint64>(sysInfo->cached_pages) + static_cast<uint64>(sysInfo->block_cache_pages)) * B_PAGE_SIZE;
 	BString cachedStr;
 	::FormatBytes(cachedStr, cachedBytes);
 	BString cachedUsage;
